Make front and back lidar Make DriveByDistance backwards Make intake belt run for autonomous Optimize DriveByDistance speed Make DriveStraight have overrideable getVelocity() method DriveToWall stopping distance adjustment Make LidarStop mostly work (does not detect stop distance properly) Add ReduceOneBall Add rumbling SpinnerLiftUp adjust speed

+++ src/main/java/frc/robot/XBox.java
@@ -67,11 +67,11 @@ public class XBox {
 
   }
 
-  public void setLeftRumble(int intensity) {
+  public void setLeftRumble(double intensity) {
     controller.setRumble(RumbleType.kLeftRumble, intensity);
   }
 
-  public void setRightRumble(int intensity) {
+  public void setRightRumble(double intensity) {
     controller.setRumble(RumbleType.kRightRumble, intensity);
   }
 }
